backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
yfinance>=1.2.0
pandas==2.2.0
numpy==1.26.3
pyyaml==6.0.1
python-dateutil==2.8.2
pydantic==2.5.3
requests==2.31.0
alpha-vantage==2.3.1
twelvedata==1.2.28
fmp-python==0.1.5
apscheduler==3.10.4
